ExoCcycleGeo v18.11.2
Cleaned up some code that was no longer useful. Still going back and
forth on the -fixed_pressure and -fixed_volume options for the GAS_PHASE
block in the PHREEQC input file for ocean-atmosphere equilibria. The
former should be appropriate in closed systems, which is not really what
an atmosphere is, but turns out to better conserve atmospheric N which
is not modified by C cycle processes in the model. However, once C cycle
processes are turned on, the conservation is not as good as
-fixed_volume even though the code runs for longer before crashing.