ExoCcycleGeo v20.5.2
Made changes to planet thermal evolution, most notably by adding a
computation of the starting temperature from accretion (varying this
matters for the first 2-3 Gyr of evolution but eventually gets erased by
radiogenic heating) and correcting a factor of 4 mistake in the overly
simplified thermal evolution equation. Benchmark to Earth is now done
with time evolution, enabling reliable computation of outgassing as a
function of planet size and age.