ExoCcycleGeo v22.2.17
Reverted to dtime timescale of ocean-seafloor interaction, rather than
the circulation timescale tcirc ~ 10 Myr. In reality:
1- A subset dtime/tcirc of the ocean circulates through the crust during
dtime
2- That subset reacts with seafloor rock at W:R ~ 1
3- The resulting solution mixes back with the ocean (as a hydrothermal
plume)
4- Any supersaturation in the ocean from river and hydrothermal inputs
results in extra mineral precipitation.

The above steps 3 and 4 could be added to the model (and steps 1-3
skipped if no plate tectonics), but for now, in the model:
1- The full ocean circulates through the crust during dtime, at W:R ~
1*tcirc/dtime
2- The reaction leads to precipitation due to both existing
supersaturation and ocean-seafloor interaction.

This approximation is probably OK because, in either case, this is a
calculation of chemical equilibrium between the full ocean and seafloor
in proportions tcirc/dtime*W:R. However, mineral precipitation is more
likely at W:R~1.

Previously, the full ocean circulated through the crust and reacted at
W:R~1, which led to a spuriously much larger control of ocean
composition and redox by hydrothermalism.

An automated switch between plate tectonics and stagnant lid effects on
farc and the choice of PHREEQC equilibrium phases template has been
added.