ExoCcycleGeo v22.3.3
Reverted to using dtime rather than tcirc as the correct time interval
for seafloor weathering flux calculation. The flux is essentially
independent of dtime because deltaCreac, for an ocean saturated in C
before and after reaction, is essentially proportional to Mriver, i.e.,
dtime. There is a slight dependence on dtime if there was no saturation
before reaction and because the solubility of C changes slightly with
slight changes in ocean pressure, temperature, and composition, but it
should be acceptable.