ExoCcycleGeo v22.6.1
Added feedback of haze particle deposition in case the CH4 mixing ratio
exceeds 1.0 times the CO2 mixing ratio (Haqq-Misra et al. 2008,
https://doi.org/10.1089/ast.2007.0197).

In that case the CH4 mixing ratio is set to the CO2 mixing ratio, the
excedent carbon goes to a growing surface layer of haze fallout, and
atmospheric-ocean equilibrium is recalculated. If the cumulative haze
deposit exceeds 1 m, the efficiency of continental weathering is
decreased by the haze fallout thickness.

The conversion between mixing ratio and fallout thickness assumes that
haze material has the molar volume of KerogenC128 (Helgeson et al. 2009,
https://doi.org/10.1016/j.gca.2008.03.004).

The carbon in haze is assumed to instantaneously drop to the surface
because even on Titan, with CH4/CO2 >> 1, the mass of carbon in
atmospheric methane is much greater than that in atmospheric haze.