ExoCcycleGeo v22.10.1
1. Changes in handling of ocean redox after seafloor weathering from
June '22 that may need to be reverted.

2. Fixed typo ("macos64" -> "macosx64") in calling alphaMELTS that was
causing that system call to be overlooked. When called every step, the
duration and outcome of alphaMELTS calculations is much less
predictable. Auto-interrupting the computation after 30 seconds does not
fix this issue as seemed to be the case back when this was coded last
winter. Instead, that typo caused bypassing of alphaMELTS and
interpolation of the P-T profile of a previous aMELTS output with the
geotherm at each timestep to compute the amount of melt. While
increasingly inaccurate away from the initial aMELTS P-T profile, having
a few aMELTS outputs on hand and deciding which to use based on
whichever the geotherm is closest to could partially fix the issue.
Looking into this "lookup table" approach next, because calling
alphaMELTS at each timestep is clearly impractical.