ExoCcycleGeo v23.8.2
Thoroughly revised the geodynamic code to remove any use or mention of
the brittle-ductile transition and instead incorporate the scaling of
Solomatov (1995) in computing heat transfer in the plate tectonics and
stagnant lid regimes. The choice of tectonic regime is now a user input.

Outgassing (to be revised) and weathering (to be left as is) are
currently commented out.

The next step is to validate or tune the geodynamic model to modern
Earth.