ExoCcycleGeo v23.9.3
Thoroughly reworked how outgassing is computed in plate tectonics mode.
Unlike in stagnant lid mode, melt is only computed at one near-surface
depth, and the volume of melted material is estimated as the {ridge+rift
surface area} * {convective velocity} * {melt fraction}.

The ridge+rift length varies with the Ra number as done previously for
seafloor weathering, but the present-day length has been increased from
about 65000 km to about 87000 km to account for continental rifts as
well. Maybe the ridge length alone should be scaled back for seafloor
weathering with (1 - {land fraction}).

The ridge+rift width is set to 50 km, constant, although it could be
made to vary with convective velocity.

The resulting outgassing flux matches present-day estimates for Earth.

Now, there is an unexplained factor of 3.6 increase in outgassing flux
in stagnant lid mode compared to v23.9.2, to be reconciled.